+++ test/inherited_provider_test.dart
@@ -18,6 +18,25 @@ BuildContext get context => find.byType(Context).evaluate().single;
 T of<T>([BuildContext c]) => Provider.of<T>(c ?? context, listen: false);
 
 void main() {
+  testWidgets('Provider.of(listen: false) outside of build works when it loads a provider', (tester) async {
+    final notifier = ValueNotifier(42);
+    await tester.pumpWidget(
+      MultiProvider(
+        providers: [
+          ValueListenableProvider.value(value: notifier),
+          ProxyProvider<int, String>(update: (a, b, c) => '$b'),
+        ],
+        child: Context(),
+      ),
+    );
+
+    expect(Provider.of<String>(context, listen: false), '42');
+
+    notifier.value = 21;
+    await tester.pump();
+
+    expect(Provider.of<String>(context, listen: false), '21');
+  });
   testWidgets('new value is available in didChangeDependencies', (tester) async {
     final didChangeDependencies = ValueBuilderMock<int>();
     final build = ValueBuilderMock<int>();
@@ -153,7 +172,26 @@ void main() {
 
     expect(
       () => Provider.of<int>(context),
-      throwsAssertionError,
+      throwsA(
+        isA<AssertionError>().having(
+          (source) => source.toString(),
+          'toString',
+          endsWith('''
+Tried to listen to a value exposed with provider, from outside of the widget tree.
+
+This is likely caused by an event handler (like a button's onPressed) that called
+Provider.of without passing `listen: false`.
+
+To fix, write:
+Provider.of<int>(context, listen: false);
+
+It is unsupported because may pointlessly rebuild the widget associated to the
+event handler, when the widget tree doesn't care about the value.
+
+The context used was: Context
+'''),
+        ),
+      ),
     );
 
     expect(Provider.of<int>(context, listen: false), equals(42));