+++ CHANGELOG.md
@@ -67,6 +67,11 @@
 * Added a `Locator` typedef and an extension on [BuildContext], to help with
   being able to read providers from a class that doesn't depend on Flutter.
 
+# 4.0.3
+
+- improved error message when `Provider.of` is called without specifying
+  `listen: false` outside of the widget tree.
+
 # 4.0.2
 
 - fix `Provider.of` returning the previous value instead of the new value